# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Riser API",
        "description": "API to interact with Riser",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Dynamic host server"
        }
    ],
    "paths": {
...
        "/api/v1/devices": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Store Device token for user",
                "description": "Store Device token for user",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Device token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete Device token for user",
                "description": "Delete Device token for user",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Device token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer Token",
                "scheme": "bearer"
            }
        }
    }
}